feat(os): pick GitHub repos via installation Octokit list
Reuse listReposAccessibleToInstallation in the repo IDE link form
so you filter and click instead of typing owner/repo. Shared picker
with the repos-page "Add from GitHub" wizard.
